+++ src/main/java/nl/quintor/recipe/ingredient/IngredientRepository.java
@@ -3,8 +3,6 @@ package nl.quintor.recipe.ingredient;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Set;
-
 @Repository
-public interface IngredientRepository extends CrudRepository<Ingredient, Long> {
+public interface IngredientRepository extends CrudRepository<Ingredient, Integer> {
 }
