+++ src/main/java/nl/quintor/recipe/recipe/RecipeRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
  * Handles persistence for recipe
  */
 @Repository
-public interface RecipeRepository extends CrudRepository<Recipe, Long> {
+public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
 }
